requirements.txt
streamlit
pymupdf
faiss-cpu
transformers
sentence-transformers
numpy
pandas
structlog
trafilatura
langchain
torch
fast_graphrag
google-generativeai
langchain_community
openai
unstructured
InstructorEmbedding
chardet
langchain_huggingface
nest-asyncio 
PyPDF2
langchain-community
tiktoken
langchain-openai
google-generativeai
lightrag-hku
pipmaster
networkx
nano_vectordb
trafilatura
pdfplumber